refactor: 북마크 confirm·chapter-delete 모달을 bookmark-modals.js로 분리 (ADR-034 후속 PR5a)
모달 다이얼로그 분리 1단계. confirm(범용 삭제 확인) + chapter-delete(장
북마크 토글-오프 피커)를 신규 bookmark-modals.js로 이동. 두 모달의 overlay·
DOM ref·정적 리스너·Escape 우선순위를 모듈이 소유한다.

모달→렌더 순환을 의존성 주입으로 끊음: bookmark.js 가 시작 시
initBookmarkModals({ rerenderActiveBookmarkTree, refreshBookmarkHeaderBtn })
로 콜백을 주입, 모달은 bookmark.js 를 import 하지 않는다(순환 import 없음 →
tsc 가 못 잡는 평가시점 깨짐 구조적 차단).

Escape 스택은 모듈의 closeTopmostModal() 로 옮기고 bookmark.js 라우터가
위임. route() 가 쓰던 window.close{Confirm,ChapterDelete}Modal facade 는
appOverlay.closeAllOverlays() 로 이미 대체된 잔재라 주석으로 명시(후속 정리
대상). bookmark.js 3044 -> 2916 줄, bookmark-modals.js 신규 209 줄.
