+++ js/app/bookmark.js
@@ -44,6 +44,16 @@ import {
   _isActiveBookmark, _hasActiveDescendant, setRenderPathname,
 } from "./bookmark-core.js";
 
+// Modal dialogs split out to bookmark-modals.js (ADR-034 후속 PR5). They own
+// their overlays + DOM refs + Escape stack (closeTopmostModal); bookmark.js
+// injects its render callbacks once via initBookmarkModals() so the modal→render
+// path needs no circular import back here.
+import {
+  initBookmarkModals, closeTopmostModal,
+  openConfirmModal, closeConfirmModal,
+  openChapterDeleteModal, closeChapterDeleteModal,
+} from "./bookmark-modals.js";
+
 
 // ── BEGIN DRAG_CORE ──
 // Exercised by tests/unit/bookmark.test.js. The test loader concatenates
@@ -540,18 +550,7 @@ const $bmMergeBody = _$("bm-merge-body");
 const $bmMergeYes = _$("bm-merge-yes");
 const $bmMergeNo = _$("bm-merge-no");
 const $bmMergeCancel = _$("bm-merge-cancel");
-const $bmConfirmScrim = _$("bm-confirm-scrim");
-const $bmConfirmModal = _$("bm-confirm-modal");
-const $bmConfirmTitle = _$("bm-confirm-title");
-const $bmConfirmBody = _$("bm-confirm-body");
-const $bmConfirmOk = _$("bm-confirm-ok");
-const $bmConfirmCancel = _$("bm-confirm-cancel");
-const $bmChapterDeleteScrim = _$("bm-chapter-delete-scrim");
-const $bmChapterDeleteModal = _$("bm-chapter-delete-modal");
-const $bmChapterDeleteAll = /** @type {HTMLInputElement} */ (_$("bm-chapter-delete-all"));
-const $bmChapterDeleteList = _$("bm-chapter-delete-list");
-const $bmChapterDeleteConfirm = /** @type {HTMLButtonElement} */ (_$("bm-chapter-delete-confirm"));
-const $bmChapterDeleteCancel = _$("bm-chapter-delete-cancel");
+// $bmConfirm* / $bmChapterDelete* refs moved to bookmark-modals.js (PR5a).
 const $verseSelectBar = _$("verse-select-bar");
 const $verseSelectCount = _$("verse-select-count");
 const $verseSelectBookmarkBtn = /** @type {HTMLButtonElement} */ (_$("verse-select-bookmark-btn"));
@@ -2086,133 +2085,7 @@ function openMergeDialog(candidates, incomingSpec, mode, fallbackContext = null)
 }
 
 // ── Destructive confirm modal ──
-// Reusable confirmation for destructive actions (bookmark/folder delete,
-// header bookmark toggle-off). Replaces the old native window.confirm() so the
-// prompt is themed, focus-trapped, and stacks above the drawer/save modals.
-/**
- * @param {{ title: string, message: string, confirmLabel?: string, onConfirm: () => void }} opts
- */
-// Overlay lifecycle (scrim/hidden/focus-trap/focus-restore) is owned by the
-// shared controller (ADR-032). closeOnEsc stays off: this modal participates in
-// bookmark.js's stacked Escape router (confirm > chapter-delete > … > drawer),
-// which calls closeConfirmModal() for the topmost overlay. A controller-level
-// Escape listener here would double-handle and also close whatever sits beneath.
-// Default initial focus = cancel (safe action); destructive button is opt-in.
-const confirmOverlay = createOverlay({
-  panel: $bmConfirmModal,
-  scrim: $bmConfirmScrim,
-  initialFocus: () => $bmConfirmCancel,
-  onClose: () => { $bmConfirmOk.onclick = null; },
-});
-
-function openConfirmModal({ title, message, confirmLabel = "삭제", onConfirm }) {
-  $bmConfirmTitle.textContent = title;
-  $bmConfirmBody.textContent = message;
-  $bmConfirmOk.textContent = confirmLabel;
-  $bmConfirmOk.onclick = () => {
-    closeConfirmModal();
-    onConfirm();
-  };
-  confirmOverlay.open();
-}
-
-function closeConfirmModal() {
-  confirmOverlay.close();
-}
-
-// Header bookmark toggle-off (mobile): instead of a blunt "delete all", present
-// the chapter's bookmarks (whole-chapter + verse ranges) as a checkbox list so
-// the reader removes only what they mean to. "전체 선택" ticks them all at once.
-// Nothing is pre-selected (the select-all is the bulk affordance) and the
-// delete button stays disabled until at least one row is ticked.
-/**
- * @param {BookmarkTreeBookmark[]} candidates
- */
-// closeOnEsc off: in the stacked Escape router. onClose clears per-open
-// handlers; returnFocus restores the pre-open focus (ADR-032).
-const chapterDeleteOverlay = createOverlay({
-  panel: $bmChapterDeleteModal,
-  scrim: $bmChapterDeleteScrim,
-  initialFocus: () => $bmChapterDeleteCancel,
-  onClose: () => {
-    $bmChapterDeleteConfirm.onclick = null;
-    $bmChapterDeleteCancel.onclick = null;
-    $bmChapterDeleteAll.onchange = null;
-  },
-});
-
-function openChapterDeleteModal(candidates) {
-  if (!candidates.length) return;
-  /** @type {Set<string>} */
-  const selected = new Set();
-
-  const syncChrome = () => {
-    const state = _selectAllState(selected.size, candidates.length);
-    $bmChapterDeleteAll.checked = state === "all";
-    $bmChapterDeleteAll.indeterminate = state === "some";
-    $bmChapterDeleteConfirm.textContent = _deleteBtnLabel(selected.size);
-    $bmChapterDeleteConfirm.disabled = selected.size === 0;
-  };
-
-  clearNode($bmChapterDeleteList);
-  /** @type {HTMLInputElement[]} */
-  const rowChecks = [];
-  candidates.forEach((bm, i) => {
-    const id = `bm-chapter-del-${i}`;
-    const li = el("li", { className: "bm-chapter-delete-item" });
-    const labelEl = el("label", { className: "bm-chapter-delete-label", for: id });
-    const input = /** @type {HTMLInputElement} */ (el("input", { type: "checkbox", id }));
-    rowChecks.push(input);
-    input.addEventListener("change", () => {
-      if (input.checked) selected.add(bm.id);
-      else selected.delete(bm.id);
-      syncChrome();
-    });
-    const book = (window.getBooksCache() ?? []).find((b) => b.id === bm.bookId);
-    const bookName = book ? (book.short_name_ko || book.name_ko) : bm.bookId;
-    const refText = bm.verseSpec === "all"
-      ? `${bookName} ${bm.chapter}${chUnit(bm.bookId)}`
-      : `${bookName} ${bm.chapter}:${bm.verseSpec}`;
-    const text = el("span", { className: "bm-chapter-delete-text" });
-    text.appendChild(el("span", { className: "bm-chapter-delete-item-label" }, bm.label));
-    text.appendChild(el("span", { className: "bm-chapter-delete-item-ref" }, refText));
-    labelEl.appendChild(input);
-    labelEl.appendChild(text);
-    li.appendChild(labelEl);
-    $bmChapterDeleteList.appendChild(li);
-  });
-
-  $bmChapterDeleteAll.onchange = () => {
-    const checkAll = $bmChapterDeleteAll.checked;
-    selected.clear();
-    if (checkAll) candidates.forEach((bm) => selected.add(bm.id));
-    rowChecks.forEach((c) => { c.checked = checkAll; });
-    syncChrome();
-  };
-
-  syncChrome();
-  chapterDeleteOverlay.open();
-
-  $bmChapterDeleteConfirm.onclick = () => {
-    if (!selected.size) return;
-    const store = loadBookmarks();
-    for (const bmId of selected) {
-      // Mirror swipe-row/folder delete: clear the per-device viewed timestamp
-      // so removed ids don't accrue as stale entries in the viewed map.
-      _forgetViewed(bmId);
-      removeItemById(store, bmId);
-    }
-    const removed = selected.size;
-    saveBookmarks(store);
-    _rerenderActiveBookmarkTree();
-    refreshBookmarkHeaderBtn();
-    announce(removed === 1 ? "북마크를 삭제했습니다." : `북마크 ${removed}개를 삭제했습니다.`);
-    closeChapterDeleteModal();
-  };
-  $bmChapterDeleteCancel.onclick = closeChapterDeleteModal;
-}
-
-function closeChapterDeleteModal() { chapterDeleteOverlay.close(); }
+// openConfirmModal / openChapterDeleteModal moved to bookmark-modals.js (PR5a).
 
 // ── Bookmark select mode (ADR-029 개정 / ADR-010) ──
 // In-place multi-select over the mobile /bookmarks full view, replacing the old
@@ -2803,10 +2676,13 @@ $bookmarkScrim.addEventListener("click", closeBookmarkDrawer);
 $bmSaveClose.addEventListener("click", closeSaveModal);
 $bmSaveScrim.addEventListener("click", closeSaveModal);
 
-$bmConfirmCancel.addEventListener("click", closeConfirmModal);
-$bmConfirmScrim.addEventListener("click", closeConfirmModal);
-
-$bmChapterDeleteScrim.addEventListener("click", closeChapterDeleteModal);
+// bookmark-modals.js owns its own scrim/cancel listeners; bookmark.js only
+// injects the render callbacks a modal needs to refresh the tree/header after a
+// mutation (의존성 주입 — breaks the modal→render cycle without a circular import).
+initBookmarkModals({
+  rerenderActiveBookmarkTree: _rerenderActiveBookmarkTree,
+  refreshBookmarkHeaderBtn,
+});
 
 $bmNewFolderClose.addEventListener("click", closeNewFolderModal);
 $bmNewFolderScrim.addEventListener("click", closeNewFolderModal);
@@ -2919,8 +2795,7 @@ $bmImportScrim.addEventListener("click", closeImportModal);
 document.addEventListener("keydown", (e) => {
   if (e.key === "Escape") {
     if (!$bmNewFolderModal.hidden) { e.preventDefault(); e.stopPropagation(); closeNewFolderModal(); return; }
-    if (!$bmConfirmModal.hidden) { closeConfirmModal(); return; }
-    if (!$bmChapterDeleteModal.hidden) { closeChapterDeleteModal(); return; }
+    if (closeTopmostModal()) return;  // confirm > chapter-delete (bookmark-modals.js)
     if (!$bmMoveModal.hidden) { closeMoveModal(); return; }
     if (!$bmImportModal.hidden) { closeImportModal(); return; }
     if (!$bmMergeModal.hidden) { closeMergeModal(); return; }
@@ -3004,13 +2879,10 @@ window.closeSaveModal = closeSaveModal;
 window.closeNewFolderModal = closeNewFolderModal;
 window.closeMergeModal = closeMergeModal;
 window.closeImportModal = closeImportModal;
-// Exposed so route() can dismiss the destructive-confirm overlay on any nav
-// (e.g. OS back gesture mid-confirm) — its scrim would otherwise persist over
-// the rebuilt view. Safe to call when already hidden (self-guards).
-window.closeConfirmModal = closeConfirmModal;
-// Same rationale for the chapter-delete picker (header bookmark toggle-off).
-window.closeChapterDeleteModal = closeChapterDeleteModal;
-// Same rationale for the move-to-folder modal (선택 모드 → 이동).
+// Exposed so route() can dismiss the move-to-folder overlay on any nav (e.g. OS
+// back gesture mid-move) — its scrim would otherwise persist over the rebuilt
+// view. Safe to call when already hidden (self-guards). The confirm /
+// chapter-delete equivalents moved with their modals to bookmark-modals.js.
 window.closeMoveModal = closeMoveModal;
 window.renderBookmarkTree = renderBookmarkTree;
 // Re-render whichever bookmark surface is mounted (drawer OR /bookmarks full
